+++ src/components/ui/index.ts
@@ -5,3 +5,4 @@ export { default as Input } from "./Input";
 export { default as Modal } from "./Modal";
 export { default as Textarea } from "./Textarea";
 export { default as Tooltip } from "./Tooltip";
+export { default as InfoTip } from "./InfoTip";